+++ ui/src/dashboards/constants/index.js
@@ -79,35 +79,29 @@ export const applyMasks = query => {
   const maskForWholeTemplates = '😸$1😸'
   return query.replace(matchWholeTemplates, maskForWholeTemplates)
 }
-
 export const insertTempVar = (query, tempVar) => {
   return query.replace(MATCH_INCOMPLETE_TEMPLATES, tempVar)
 }
-
 export const unMask = query => {
   return query.replace(/😸/g, ':')
 }
-
 export const removeUnselectedTemplateValues = templates => {
   return templates.map(template => {
     const selectedValues = template.values.filter(value => value.selected)
     return {...template, values: selectedValues}
   })
 }
-
 export const DISPLAY_OPTIONS = {
   LINEAR: 'linear',
   LOG: 'log',
   BASE_2: '2',
   BASE_10: '10',
 }
-
 export const TOOLTIP_CONTENT = {
   FORMAT:
     '<p><strong>K/M/B</strong> = Thousand / Million / Billion<br/><strong>K/M/G</strong> = Kilo / Mega / Giga </p>',
 }
-
 export const TYPE_QUERY_CONFIG = 'queryConfig'
+export const TYPE_SHIFTED = 'shifted queryConfig'
 export const TYPE_IFQL = 'ifql'
-
 export const DASHBOARD_NAME_MAX_LENGTH = 50