+++ ui/src/data_explorer/components/GroupByTimeDropdown.js
@@ -7,8 +7,6 @@ import Dropdown from 'shared/components/Dropdown'
 
 import {AUTO_GROUP_BY} from 'shared/constants'
 
-const {func, string, shape} = PropTypes
-
 const isInRuleBuilder = pathname => pathname.includes('alert-rules')
 const isInDataExplorer = pathname => pathname.includes('data-explorer')
 
@@ -36,6 +34,8 @@ const GroupByTimeDropdown = ({
       selected={selected || 'Time'}
     />
   </div>
+
+const {func, string, shape} = PropTypes
 
 GroupByTimeDropdown.propTypes = {
   location: shape({